backend/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
whisper>=1.1.10
ollama>=0.3.0
pydantic>=2.10.0
python-dotenv>=1.0.0
anthropic>=0.40.0
httpx>=0.27.0
